Update footer & file structure
