--- tuandat-360review/src/container/footer/Footer.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import Detail from '../../components/FooterDetail/detail.js'
-
-const footer = () => {
-    <Detail />
-}
-
-export default footer;
