+++ tuandat-360review/src/container/footer/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Detail from '../../components/FooterDetail';
+
+const footer = () => (
+    <Detail />
+)
+
+export default footer;
